+++ packages/orq-rc/src/models/operations/listdatasources.ts
@@ -176,7 +176,7 @@ export const ListDatasourcesData$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KR3N6V46XJQ6S1YP1VKQF8H9"),
+  _id: z.string().default("01KR448YCP2AYRNVNV62YMW3VH"),
   display_name: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$inboundSchema,