+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3628,7 +3628,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KR3N6T87XWTRK41KXZW7P0MM"),
+  id: z.string().default("01KR448XFMPYQ34B96QZ7MWRD6"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),