+++ packages/orq-rc/RELEASES.md
@@ -8969,3 +8969,13 @@ Based on:
 - [typescript v4.9.0-rc.35] packages/orq-rc
 ### Releases
 - [NPM v4.9.0-rc.35] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.35 - packages/orq-rc
+
+## 2026-05-08 15:49:15
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.9.0-rc.36] packages/orq-rc
+### Releases
+- [NPM v4.9.0-rc.36] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.36 - packages/orq-rc